Island not in connection, fix self connect.
The island is not a property of a connection. For now there will be only
one island per diffuser. If you want to run two diffused applications,
maybe you can run two diffusers.

Self connect would cause problems as both the client and the server side
of the connection would use the same slot in the data structure used to
store the connection. Hacked a solution by using a `self` index for the
`from` address when self-connecting.