Debug messages, propagate buckets to regsitrar.
Leaving some debug messages in place that I keep adding and deleting.
We're not done yet.

The registrar was not getting the bucket count so it was not correctly
updating peers with registered sockets.

Found a bug in Staccato the pursuit of which added a lot of debugging
noise, most of which has been removed.