Sometimes empty after waking laptop from sleep.
The list of hosts from mingle is sometimes null when waking laptop from
sleep. What is the difference though between a lookup failure and an
affirmative lookup of zero? How are we supposed to handle the absence of
coarse discovery? It seems like there should be a timeout when we've
lost confidence in our environment, but we shouldn't throw in the towel
on the first failure. We shouldn't plug along indefinately either.