Assert contiguous packets, short-circuit to self.
Add a series number to the envelope and ensure that the series is
correct on the other side of the connection. If this doesn't work we
know that we have a problem in Conduit Window. We depend on receiving
all packets in order.

When sending messages to self just dump them into the inbox, do not send
over the network.