Correctly select client among peer workers.
Their is one connection between each individual worker process.
Previously the code was selecting who would initiate the connection --
the client -- by selecting the instance with the lesser promise as the
client. Dialing up the workers means that all the workers peers are
clients of each other, so we'd end up opening a connection each way and
then loose track of the connection as the server side over wrote the
client side in the connection collection. The client is now the worker
with the lesser promise or the lesser index if the promises are equal.