Slam socket shut on socket destruct.
Otherwise it hangs and we have Window problems. Not sure why we have
Window problems. I want to return to this and remove the socket destroy
in order to debug Window. Then I'd like to sort out how to shutdown one
of these sockets. Maybe `destroy` is the right way to do it, but I
imagined I'd let it drain correctly.