Window debugging funnel.
Created a logging funnel to debug Conduit Window and found that because
it was relying on the outbox provided to connect to maintain its
reservoir of messages to resent on reconnect, it would lose messages
when the reservoir was truncated should the `end` method of the outbox
be called by provider.

Created a debugging funnel as an service with a single Inlet based
listener. Rather than spewing complete log lines it prints the major
events of create and connecting, and then gathers logs lines from the
Windows to be spewed if the receiving end of a Window gets ahead of the
sending end. This revealed the above issue.

Windows now have an id that they can report. I exposed the Paxos
republic to differentiate Windows between consensus restarts. A Window
is defined by the promise and index of both the to and from direction
plus the republic.

Also, closing sockets send by the Diffuser constituent process until we
receive our first route. We don't know our promise and index until we
get a route, so we can't label our Windows correctly.

Added the ability to patch `node_modules` during a docker build. You
generate a `node_modules.patch` in the root directory from your
`node_modules` with your hacks, and a `_node_modules` copy of the
original `node_modules`.

Olio updates where made to hopefully shutdown a cluster child correctly.
It appears that calling disconnect creates a new variation of shutdown
behavior that waits for servers to close, but not for the entire Node.js
work queue to unwind. This defeats Destructible. Olio now has its own
shutdown IPC message to shutdown servers. Closing a server explicitly
will remove it from the parent's round-robin. I know this from reading
the Node.js code. We appear to have Destructible and scram working
again.

Added a wait for close in the Diffuser constituent.

Made the chaos monkey optional. Specified in the construction of a
`Diffuser` object in a client program. Would rather it was specified in
the construction of the Diffuser constituent, but for now I just want to
turn off the code in production without commenting out code.