Use HTTP in dummy application.
I'd created a TCP server that reads and writes streaming JSON as the
ingress for the dummy application, however Diffuser is the first
application I'm writing that uses TCP as it's network protocol and I'm
working through my misunderstandings of TCP through Diffuser, and not
through the dummy application. My understanding of HTTP is more mature.
I'm able to stop servers and clients without hanging. I'm able to close
the connections cleanly. Therefore I've changed the ingress protocol to
HTTP and now worker is not crashing for its own reasons.