Added series assertions, but they're asserting.
Added series assertions and they are working, but they are asserting.
The integration-ish test is not passing. Checking in at this point so
that I don't break what seems to be working as I fix what is not
working. If series are indeed off, then it is time to inspect Conduit
and see if it is somehow dropping packets.