Fire-and-forget routing, multiple messages.
Moving toward a different interface where you can send messages for
multiple entries at once and ignore the response. The response is still
gathered for the sake of logging response times. The `route2` method
accepts a destiation, either a registered receiver or a router, either a
message or an array of messages, an optional logging context message,
and an optional callback. The context message is used by our dear user
to log timings according and tag them with tags relevent to our dear
user's application.