Working Route2, add `via` to messages.
Errors galore as I was using `from` to track the series of the messages
in the dispatcher, a series number for each incoming socket, but I was
forwarding messages using the `from` of the original message so that the
message comming off the socket could not be `from` the actual other side
of the socket, but instead a forwarded message.

The new route is now working correctly. Had to ensure that I was
propgating both the packet version number and the request index.